+++ core/src/main/java/com/gabcode/core/domain/model/Item.kt
@@ -5,12 +5,12 @@ import com.gabcode.core.extension.empty
 data class Item(
     val id: String,
     val title: String,
-    val price: Int = 0,
+    val price: Double = 0.0,
     val thumbnail: String = String.empty(),
     val pictures: List<Picture> = listOf()
 ) {
     companion object {
-        fun empty() = Item(String.empty(), String.empty(), 0, String.empty(), listOf())
+        fun empty() = Item(String.empty(), String.empty(), 0.0, String.empty(), listOf())
     }
 }
